+++ src/sw_server.h
@@ -36,7 +36,7 @@ struct sw_server_input {
   float mismatch;   /**< Penalty for mismatch. */
   float gap_open;   /**< Penalty for gap opening. */
   float gap_extend; /**< Penalty for gap extending. */
-  float min_score;  /**< Minimum normalized score (0..1) for valid alignments. */
+  int   min_score;  /**< Minimum normalized score (0..100) for valid alignments. */
   unsigned int flank_length; /**< Length to extend the CAL region. */
   unsigned int write_size;   /**< Size of the writing batch (to disk). */
   
@@ -95,7 +95,7 @@ void sw_optarg_init(float gap_open, float gap_extend,
  */
 void sw_server_input_init(list_t* sw_list_p, list_t* write_list_p, unsigned int write_size, 
 			  float match, float mismatch, float gap_open, float gap_extend, 
-			  float min_score, unsigned int flank_length, genome_t* genome_p,
+			  int min_score, unsigned int flank_length, genome_t* genome_p,
 			  size_t max_intron_size, int min_intron_size, size_t seed_max_distance, 
 			  bwt_optarg_t* bwt_optarg_p, avls_list_t *avls_list,
 			  cal_optarg_t *cal_optarg_p, bwt_index_t *bwt_index_p,
